+++ crates/djmxcreation-backend-axum/src/router/project_router.rs
@@ -27,8 +27,8 @@ pub struct Params {
 
 #[derive(Deserialize)]
 pub struct PaginationQueryParams {
-    pub page: i32,
-    pub size: i32,
+    pub page: i64,
+    pub size: i64,
     pub adult: Option<bool>,
     pub visible: bool,
 }
